requirements.txt
anthropic
pydantic==2.10.4
click==8.1.8
python-dotenv==1.0.1